test(app): share throwing storage test helper across theme suites (#102)
