+++ app/composables/__tests__/useTheme.test.ts
@@ -1,29 +1,10 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { useTheme } from '@/composables/useTheme';
 import { THEME_STORAGE_KEY } from '@/utils/theme';
+import { installThrowingStorageStub } from '#tests/test-helpers/storageStub';
 describe('useTheme', () => {
-  let originalLocalStorage: PropertyDescriptor | undefined;
-  const stubThrowingStorage = () => {
-    Object.defineProperty(window, 'localStorage', {
-      configurable: true,
-      value: {
-        get length() {
-          return 0;
-        },
-        clear() {},
-        getItem(): string | null {
-          throw new Error('SecurityError: Access is denied for this document');
-        },
-        key: () => null,
-        removeItem() {},
-        setItem(): void {
-          throw new Error('SecurityError: Access is denied for this document');
-        },
-      } satisfies Storage,
-    });
-  };
+  let restoreStorage: (() => void) | undefined;
   beforeEach(() => {
-    originalLocalStorage = Object.getOwnPropertyDescriptor(window, 'localStorage');
     vi.restoreAllMocks();
     window.localStorage.clear();
     document.documentElement.removeAttribute('data-theme');
@@ -31,9 +12,8 @@ describe('useTheme', () => {
     clearNuxtState(['theme-mode', 'theme-hydrated']);
   });
   afterEach(() => {
-    if (originalLocalStorage) {
-      Object.defineProperty(window, 'localStorage', originalLocalStorage);
-    }
+    restoreStorage?.();
+    restoreStorage = undefined;
     vi.restoreAllMocks();
   });
   it('defaults to dark when nothing is persisted', () => {
@@ -85,7 +65,7 @@ describe('useTheme', () => {
   });
   it('preserves in-memory mode across multiple useTheme calls when storage is unavailable', () => {
     const { setThemeMode } = useTheme();
-    stubThrowingStorage();
+    restoreStorage = installThrowingStorageStub();
     setThemeMode('light');
     const anotherConsumer = useTheme();
     expect(anotherConsumer.themeMode.value).toBe('light');
